refactor(scripts): organize placement launch plumbing
