+++ dflash/scripts/bench_he.py
@@ -16,7 +16,8 @@ import sys
 import tempfile
 from pathlib import Path
 
-from placement.backend_device import apply_backend_visible_devices, TestDflashLaunchArgs
+from placement.backend_device import apply_backend_visible_devices
+from placement.test_dflash_args import TestDflashLaunchArgs
 
 
 ROOT = Path(__file__).resolve().parent.parent
