+++ dflash/scripts/placement/__init__.py
@@ -1,1 +1,1 @@
-"""Script-side backend/device placement helpers."""
+"""Script-side placement helpers."""
