+++ dflash/scripts/server.py
@@ -40,7 +40,7 @@ from _prefill_hook import (
     PrefillConfig, add_cli_flags, config_from_args,
     compress_text_via_daemon, _drain_until_sentinel,
 )
-from placement.backend_device import TestDflashLaunchArgs
+from placement.server_resolver import resolve_server_placement
 from prefix_cache import DaemonStdoutBus, PrefixCache
 from tool_memory import ToolMemory
 
@@ -2371,10 +2371,8 @@ def main():
     if args.fa_window is not None:
         os.environ["DFLASH27B_FA_WINDOW"] = str(args.fa_window)
 
-    if args.target_gpu is not None:
-        os.environ["DFLASH_TARGET_GPU"] = str(args.target_gpu)
-    if args.draft_gpu is not None:
-        os.environ["DFLASH_DRAFT_GPU"] = str(args.draft_gpu)
+    placement = resolve_server_placement(args)
+    placement.apply_env(os.environ)
 
     if args.prefill_compression != "off":
         os.environ.setdefault("DFLASH27B_LM_HEAD_FIX", "0")
@@ -2421,36 +2419,18 @@ def main():
         drafter_tokenizer = AutoTokenizer.from_pretrained(
             prefill_cfg.drafter_tokenizer_id, trust_remote_code=True)
 
-    extra_daemon_cfg = TestDflashLaunchArgs(
-        draft_feature_mirror=args.draft_feature_mirror,
-        peer_access=args.peer_access,
-    )
-    if args.target_gpus:
-        extra_daemon_cfg = TestDflashLaunchArgs(
-            draft_feature_mirror=args.draft_feature_mirror,
-            peer_access=args.peer_access,
-            target_gpus=args.target_gpus,
-            target_layer_split=args.target_layer_split,
-            # Keep sharded daemon behavior aligned with the single-GPU server path.
-            target_split_load_draft=True,
-            target_split_dflash=True,
-        )
-        # Multi-GPU daemon mode currently does not implement SNAPSHOT/RESTORE.
-        if args.prefix_cache_slots > 0 or args.prefill_cache_slots > 0:
-            print("  [cfg] target-gpus daemon mode disables prefix/full cache slots (snapshot protocol unsupported)")
-            args.prefix_cache_slots = 0
-            args.prefill_cache_slots = 0
-    extra_daemon = extra_daemon_cfg.to_cli_args()
+    if placement.cache_slots_disabled:
+        print("  [cfg] target-gpus daemon mode disables prefix/full cache slots (snapshot protocol unsupported)")
 
     app = build_app(args.target, draft, args.bin, args.budget, args.max_ctx,
                     tokenizer, stop_ids,
                     prefill_cfg=prefill_cfg if prefill_cfg.enabled else None,
                     drafter_tokenizer=drafter_tokenizer,
-                    prefix_cache_slots=args.prefix_cache_slots,
-                    prefill_cache_slots=args.prefill_cache_slots,
+                    prefix_cache_slots=placement.prefix_cache_slots,
+                    prefill_cache_slots=placement.prefill_cache_slots,
                     prefill_cache_bytes=args.prefill_cache_bytes,
                     arch=arch,
-                    extra_daemon_args=extra_daemon or None,
+                    extra_daemon_args=placement.daemon_args or None,
                     lazy_draft=args.lazy_draft,
                     verbose_daemon=args.verbose_daemon)
 
@@ -2468,6 +2448,8 @@ def main():
     print(f"  budget    = {args.budget}")
     print(f"  max_ctx   = {args.max_ctx}")
     print(f"  tokenizer = {tokenizer_id}")
+    for line in placement.log_lines():
+        print(line)
     if args.lazy_draft:
         print("  lazy_draft= ON (decode draft parked when idle)")
     if args.verbose_daemon:
